+++ src/console/controllers/FetchController.php
@@ -9,6 +9,7 @@ use craft\queue\QueueInterface;
 use percipiolondon\staff\elements\Employer;
 use percipiolondon\staff\jobs\v2\FetchEmployeesJob;
 use percipiolondon\staff\jobs\v2\FetchEmployersJob;
+use percipiolondon\staff\jobs\v2\FetchPayRunJob;
 use percipiolondon\staff\Staff;
 use yii\helpers\Console;
 use yii\queue\redis\Queue as RedisQueue;
@@ -78,6 +79,27 @@ class FetchController extends Controller
                 'employers' => Employer::findAll(),
             ],
             'description' => 'Fetching employees',
+        ]));
+
+        $this->_runQueue();
+
+        $this->stdout('' . PHP_EOL, Console::RESET);
+        $this->stdout('--------------------------------- Done fetching from Staffology' . PHP_EOL, Console::FG_CYAN);
+        $this->stdout('' . PHP_EOL, Console::RESET);
+    }
+
+    public function actionPayRun()
+    {
+        $this->stdout('' . PHP_EOL, Console::RESET);
+        $this->stdout('--------------------------------- Start fetching data from Staffology' . PHP_EOL, Console::FG_CYAN);
+        $this->stdout('' . PHP_EOL, Console::RESET);
+
+        $queue = Craft::$app->getQueue();
+        $queue->push(new FetchPayRunJob([
+            'criteria' => [
+                'employers' => Employer::findAll(),
+            ],
+            'description' => 'Fetching pay run',
         ]));
 
         $this->_runQueue();